package name was capitalized
