+++ src/main/java/com/direwolf20/mininggadgets/client/screens/ModificationTableScreen.java
@@ -4,8 +4,8 @@ import com.direwolf20.mininggadgets.MiningGadgets;
 import com.direwolf20.mininggadgets.common.containers.ModificationTableContainer;
 import com.direwolf20.mininggadgets.common.gadget.upgrade.Upgrade;
 import com.direwolf20.mininggadgets.common.network.PacketHandler;
-import com.direwolf20.mininggadgets.common.network.Packets.PacketExtractUpgrade;
-import com.direwolf20.mininggadgets.common.network.Packets.PacketInsertUpgrade;
+import com.direwolf20.mininggadgets.common.network.packets.PacketExtractUpgrade;
+import com.direwolf20.mininggadgets.common.network.packets.PacketInsertUpgrade;
 import com.mojang.blaze3d.platform.GlStateManager;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screen.Screen;
