+++ src/main/java/com/direwolf20/mininggadgets/common/events/ServerTickHandler.java
@@ -1,7 +1,7 @@
 package com.direwolf20.mininggadgets.common.events;
 
 import com.direwolf20.mininggadgets.common.network.PacketHandler;
-import com.direwolf20.mininggadgets.common.network.Packets.PacketDurabilitySync;
+import com.direwolf20.mininggadgets.common.network.packets.PacketDurabilitySync;
 import net.minecraft.util.Tuple;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
