+++ src/main/java/com/direwolf20/mininggadgets/common/network/packets/PacketDurabilitySync.java
@@ -1,4 +1,4 @@
-package com.direwolf20.mininggadgets.common.network.Packets;
+package com.direwolf20.mininggadgets.common.network.packets;
 
 import com.direwolf20.mininggadgets.common.tiles.RenderBlockTileEntity;
 import net.minecraft.client.Minecraft;
