+++ src/main/java/com/direwolf20/mininggadgets/common/network/packets/PacketInsertUpgrade.java
@@ -1,4 +1,4 @@
-package com.direwolf20.mininggadgets.common.network.Packets;
+package com.direwolf20.mininggadgets.common.network.packets;
 
 import com.direwolf20.mininggadgets.common.containers.ModificationTableCommands;
 import com.direwolf20.mininggadgets.common.containers.ModificationTableContainer;
@@ -12,34 +12,24 @@ import net.minecraftforge.fml.network.NetworkEvent;
 
 import java.util.function.Supplier;
 
-public class PacketExtractUpgrade {
+public class PacketInsertUpgrade {
 
     private final BlockPos pos;
-    private final String upgrade;
-    private final int nameLength;
-    private final boolean isShiftHeld;
 
-    public PacketExtractUpgrade(BlockPos blockPos, String upgrade, int nameLength, boolean isShiftHeld) {
+    public PacketInsertUpgrade(BlockPos blockPos) {
         this.pos = blockPos;
-        this.upgrade = upgrade;
-        this.nameLength = nameLength;
-        this.isShiftHeld = isShiftHeld;
     }
 
-    public static void encode(PacketExtractUpgrade msg, PacketBuffer buffer) {
-        buffer.writeInt(msg.nameLength);
+    public static void encode(PacketInsertUpgrade msg, PacketBuffer buffer) {
         buffer.writeBlockPos(msg.pos);
-        buffer.writeString(msg.upgrade);
-        buffer.writeBoolean(msg.isShiftHeld);
     }
 
-    public static PacketExtractUpgrade decode(PacketBuffer buffer) {
-        int strLength = buffer.readInt();
-        return new PacketExtractUpgrade(buffer.readBlockPos(), buffer.readString(strLength), strLength, buffer.readBoolean());
+    public static PacketInsertUpgrade decode(PacketBuffer buffer) {
+        return new PacketInsertUpgrade(buffer.readBlockPos());
     }
 
     public static class Handler {
-        public static void handle(PacketExtractUpgrade msg, Supplier<NetworkEvent.Context> ctx) {
+        public static void handle(PacketInsertUpgrade msg, Supplier<NetworkEvent.Context> ctx) {
             ctx.get().enqueueWork(() -> {
                 ServerPlayerEntity player = ctx.get().getSender();
                 if (player == null) return;
@@ -51,7 +41,7 @@ public class PacketExtractUpgrade {
                 if (!(te instanceof ModificationTableTileEntity)) return;
                 ModificationTableContainer container = ((ModificationTableTileEntity) te).getContainer(player);
 
-                ModificationTableCommands.extractButton(container, player, msg.upgrade, msg.isShiftHeld);
+                ModificationTableCommands.insertButton(container);
             });
 
             ctx.get().setPacketHandled(true);
